requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-auth-ldap==1.7.0
django-cors-headers==2.5.2
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
idna==2.8
isort==4.3.16
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.3.1
python-ldap==3.2.0
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.11.1
